Mick card update modal

See merge request BK93/project-truelime-frontend!18

+++ src/app/components/boards/board/board.component.ts
@@ -17,6 +17,7 @@ import { Subscription } from 'rxjs';
 import { LaneModalComponent } from './modals/lane-modal/lane-modal.component';
 import { BoardModalComponent } from './modals/board-modal/board-modal.component';
 import { Lane } from '@app/models/lane';
+import { Card } from '@app/models/card';
 
 @Component({
   selector: 'app-board',
@@ -122,8 +123,18 @@ export class BoardComponent implements OnInit, OnDestroy {
   }
 
   deleteCard(laneId: string, cardId: string) {
-    this.boardService
+    if(confirm(`Weet u zeker dat u dit kaartje wilt verwijderen?`)){
+      this.boardService
       .deleteCard(this.board.id, laneId, cardId)
+      .pipe(first())
+      .subscribe(data => {});
+  
+    }
+  }
+
+  updateCard(laneId: string, cardId: string, card: Card){
+    this.boardService
+      .updateCard(this.board.id, laneId, cardId, card)
       .pipe(first())
       .subscribe(data => {});
   }
